pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["configuration.py"]

[project]
name = "dbx-benchmarking"
version = "0.1.0"
description = "Databricks query benchmarking across compute targets, connectors, and materialization strategies"
requires-python = ">=3.11"
dependencies = [
    "databricks-connect",
    "databricks-sdk",
    "databricks-sql-connector",
    "pyarrow",
    "python-dotenv",
]
